--- frontend/src/components/app.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from "react";
-export default class App extends Component {
-  render() {
-    return (
-      <div>
-        {this.props.children}
-      </div>
-    );
-  }
-}
